requirements.txt
mistralai
python-dotenv
langchain
langchain-mistralai
langchain-community
llama-index
llama-index-llms-mistralai